+++ .agentic-workspace/planning/lanes/lane-p2-runtime-packet-owner-boundaries.lane.json
@@ -24,8 +24,8 @@
     {
       "id": "implement-packets",
       "title": "Extract implement runtime packets",
-      "status": "planned",
-      "execplan_ref": "",
+      "status": "completed",
+      "execplan_ref": ".agentic-workspace/planning/execplans/archive/p2-implement-runtime-packets.plan.json",
       "depends_on": [
         "startup-packets"
       ],
@@ -52,16 +52,16 @@
       "purpose_for_lane": "Satisfy #1773 without weakening proof or closeout blocking."
     }
   ],
-  "current_slice": "implement-packets",
+  "current_slice": "planning-packets",
   "acceptance_boundary": "The lane is complete when #1770 through #1773 each have owner-scoped runtime modules, stable compatibility imports, updated inventories, and focused tests proving unchanged emitted packet contracts.",
   "proof_strategy": "Run focused start/implement/planning/proof tests for each slice, generated command package checks when inventories change, and make test-workspace for cross-runtime extraction.",
   "proof_aggregation": {
     "status": "partial",
     "evidence": [
-      "#1770 startup-packets completed; proof is recorded in .agentic-workspace/planning/execplans/archive/p2-startup-runtime-packets.plan.json and included make test-workspace, make lint-workspace, generated command package checks, operation conformance, contract tooling, structured file inventory, ruff, maintainer surfaces, schema reference docs, AW defaults/summary/doctor, full tests/test_workspace_cli.py, generator --check, AW primitive ownership, and git diff --check."
+      "#1770 startup-packets completed; proof is recorded in .agentic-workspace/planning/execplans/archive/p2-startup-runtime-packets.plan.json and included make test-workspace, make lint-workspace, generated command package checks, operation conformance, contract tooling, structured file inventory, ruff, maintainer surfaces, schema reference docs, AW defaults/summary/doctor, full tests/test_workspace_cli.py, generator --check, AW primitive ownership, and git diff --check.",
+      "#1771 implement-packets completed; proof is recorded in .agentic-workspace/planning/execplans/archive/p2-implement-runtime-packets.plan.json and included make test-workspace, make lint-workspace, generated command package checks, operation conformance, contract tooling, structured file inventory, ruff, schema reference docs, AW defaults/summary/doctor, full tests/test_workspace_cli.py, generator --check, AW primitive ownership, and git diff --check."
     ],
     "known_gaps": [
-      "#1771 implement runtime packets",
       "#1772 Planning runtime packets",
       "#1773 proof and closeout runtime packets"
     ]
@@ -71,8 +71,8 @@
   "parent_close_permission": "do-not-close-parent",
   "closeout_state": {
     "status": "open",
-    "summary": "#1770 startup runtime packet extraction is complete in the first lane slice.",
-    "residual_work": "Implement #1771, #1772, and #1773 in order.",
+    "summary": "#1770 startup and #1771 implement runtime packet extractions are complete.",
+    "residual_work": "Implement #1772 and #1773 in order.",
     "next_owner": "current agent"
   },
   "references": [
