+++ src/agentic_workspace/contracts/workspace_runtime_primitive_families.json
@@ -411,11 +411,22 @@
           "_compact_start_proof_payload",
           "_compact_tiny_required_proof_commands",
           "_context_router_family_payload",
+          "_diagnostic_profile",
           "_emit_payload",
           "_guidance_with_cli_invoke",
           "_local_chat_checkpoint_projection",
           "_selector_first_planning_safety_gate",
           "_selector_requests",
+          "_selector_requests_architecture_principles",
+          "_selector_requests_assurance_requirements",
+          "_selector_requests_change_impact",
+          "_selector_requests_plan_delegation_packet",
+          "_selector_requests_requirement_grounding",
+          "_selector_requests_reuse_pressure",
+          "_selector_requests_routine_work_context",
+          "_selector_requests_task_contract",
+          "_selector_requests_test_strategy_check",
+          "_selector_requests_verification",
           "_select_payload_fields",
           "_start_profile_for_select",
           "_start_tiny_payload_fast",
@@ -662,6 +673,7 @@
           "_package_boundary_payload",
           "_package_json_scripts",
           "_package_scripts_without_negative_routes",
+          "_read_task_text_from_file",
           "_resolve_target_root",
           "_subsystem_matches_for_changed_paths",
           "_target_proof_capabilities",
